form component for announcement
